Klokku internal calendar implemented (#7)
This provides the functionality to store events inside the Klokku
application without the need to use external calendar service.
Additionally, it is now possible to modify the past events in the
calendar. The events modifications are done in a "sticky" way to make it
easier to keep the full calendar occupied by the events.

Related to https://github.com/klokku/klokku-ui/pull/6